OR1K: use custom lowering for DYNAMIC_STACKALLOC
OR1K ABI defines that outgoing function operands
should reside at the top of the stack, that means
that dynamically allocated stack areas has to be placed
in between the outgoing function operands and any other
stack objects.

To achieve this a ADJDYNALLOC pseudo instruction
is inserted when DYNAMIC_STACKALLOC is lowered.
This pseudo instruction holds a copy of the
stack pointer and has the frame index of the
dynamic stack allocation as output.

Later, in emitPrologue, when we know the
largest possible call frame size (i.e. maximum
outgoing function arguments on the stack),
the ADJDYNALLOC pseudo instruction will be replaced with a:
l.addi <dyn frame idx>, <copy of sp>, <max call frame>