CodeGen support for transparent_union
Currently when unions marked with the transparent_union
attribute are used as function arguments, the declared function gets
called with the union instead of the calling convention of the first field
in the transparent union as is required by the description

This tries to improve the situation by converting the
type to the first field of the union, but there are a couple of things I am
not completely satisfied with.
The thing that bugs me most is the conversion in classifyArgument in
lib/CodeGen/TargetInfo.cpp. Since this is ABI specific, every ABI
has to do the conversion, which seems sub-optimal to me.