OR1K: Realign stack on objects with alignment > stack alignment
To be able to ensure correct alignment of stack objects that
have a larger alignment than the stack alignment, the stack
needs to be realigned on function entry since there is
no guarantee that the stack pointer will be correctly aligned.

On OR1K, the ABI states that the stack alignment is 4 bytes,
so any stack object with an alignment > 4 bytes will need to
have the stack realigned.

The realignment is done by masking out the lower bits
of the stack pointer after subtracting the current frame
offset of it, thus functions which have stack
realignment requires a frame pointer in order to properly
restore the stack pointer on function exit.
Stack objects in the current frame (local objects) are
addressed with a positive offset relative to the stack pointer
and stack objects in the previous frame are addressed with a
negative offset relative to the frame pointer.

Functions that have variable sized stack objects can not use the
stack pointer to address it local objects, since the
stack pointer will be dynamically adjusted.
Those functions require a base register, to where the
stack pointer is copied before it is dynamically adjusted.
Local objects will then be addressed by using a positive offset
relative to the base register.

As a base register, r14 has been picked.
Ideally it could be dynamically allocate to any of the
callee-saved registers.
The base register is modified early in the prologue,
so it get the same special threatment as the stack pointer (r1)
and link register (r9) and is saved with a negative offset
relative to the stack pointer before it is adjusted.