OR1K: add support for GOT referencing
This works by running a machine pass that inserts
a copy from _GLOBAL_OFFSET_TABLE_-PC into a
global base reg.

The PC value is obtained by issuing a l.jal instruction
to the next instruction (after a l.nop) and copying
the value in r9 (link register).

if the function contains calls, the global base reg
is copied into r16,
since the .plt entries expects a pointer to GOT
in that register (at least for now).

Some various changes is made to otherwise
unrelated code, hi/lo() printing is now
handled by MCExpr instead of a dedicate printMethod.

The gotpchi/lo would ideally be handled by MCExpr as well,
but since the _GLOBAL_OFFSET_TABLE_-PC is not generated
as a symbol expression, but as a binary expression,
this doesn't seem possible.
A workaround where MOVHI and ORI has duplicated
entries with special printMethods is created.
The duplicates are called MOVHI_GOTPCHI and ORI_GOTPCLO.