OR1K: Remove OR1K specific VariantKind printing from MCExp
Instead handle the printing in OR1KInstPrinter.
There are several advantages with this,
the most significant is that the printing in MCExpr only covers
SymbolRefExps, but we want to also print hi/lo gotpchi/lo etc
on i.e. binary expressions.

This allows us to kill MOVHI_GOTPCHI and ORI_GOTPCLO.